greybus: kill struct gmod_cport
The contents of the gmod_cport structure were mostly fabricated,
with the cport_id field being the only one that's meaningful.

This patch gets rid of that structure.  Instead, a new typedef
is created to represent a cport id as used by Greybus, and
everywhere a struct gmod_cport had been used before gets updated
to just use the cport id instead.

A few other spots that used other types (like int) for cport
ids are updated to use the new typedef as well.

Signed-off-by: Alex Elder <elder@linaro.org>