greybus: define Greybus-over-USB transfer buffer type
A Greybus buffer contains a transfer buffer to hold data being
transferred.  For the ES-1 AP we need to insert the destination
CPort ID as a single byte value to tell the AP (or other) Bridge
where it should send the data.

This patch defines a structure to represent that, consisting of a
one-byte cport id and an arbitrary data buffer following it.  This
avoids having to use negative array indices (and their requisite
"I really mean it" comments).

Signed-off-by: Alex Elder <elder@linaro.org>