greybus: i2c-gb: add irq protocol and irqchip support
This is the crude hack of Greybus gpio irq support masquerading
as an I2C irq. By hiding the irq in the I2C protocol driver, we
avoid the current kernel limitation on handling of GPIO hot unplug.

This hack registers an irqchip with the Greybus i2c adapter that
has one source. This is intended to be used only with the December
demo. The protocol is, however, the same that is being used in
for actual GPIO interrupt support which leverages the standard
gpiolib irqchip helpers.

NOTE: DO NOT MERGE THIS TO GREYBUS MASTER!!!

Signed-off-by: Matt Porter <mporter@linaro.org>
Reviewed-by: Alex Elder <elder@linaro.org>